pythonServices/requirements.txt
pymupdf4llm>=0.0.27
pymupdf>=1.26.3
scikit-learn
nltk
pandas
xgboost
uvicorn
fastapi
chromadb
sentence-transformers
google-genai
python-dotenv
pydantic
google-cloud-texttospeech==2.27.0
pydub